DAOS-RSP vos: Block I/O object type
This patch adds the necessary required support for adding
support for block I/O type for DAOS.

Block I/O type currenty writes I/Os directly to SSDs.
Block capacity under a-keys is configurable using a
server side env variable
BIO_BLOCK_CAPACITY

This feature has also been added to daos_perf currently.
with this patch daos_perf can be used to trigger block
I/O using a cmdline as show below.
mpirun -np 1 daos_perf -a 1 -d 10 -c LARGE -P 2G -N 20G -T daos -n 4 -s
4k -A -b 16777216 -R "U;p F;p"

VOS perf cmdline
mpirun -np 1 daos_perf -a 1 -d 10 -c LARGE -P 2G -T vos -z -n 4096 -s 4k
-A -b 16777216 -R "U;p F;p"

Although majority of the changes are for VOS layer to enable this patch,
Changes include modification to other layers too.
-- srv and bio
-- daos_client
-- daos_perf

+++ src/common/tests_lib.c
@@ -35,7 +35,7 @@ dts_oid_gen(unsigned seed)
 }
 
 daos_unit_oid_t
-dts_unit_oid_gen(daos_oclass_id_t oclass, uint8_t ofeats, uint32_t shard)
+dts_unit_oid_gen(daos_oclass_id_t oclass, uint16_t ofeats, uint32_t shard)
 {
 	daos_unit_oid_t	uoid;
